+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPWM8847AF870MGEZW16RM72"
+        "01KPX4DPK381Z5Z35SVVM9JGS6"
     )
     r"""The unique identifier of the data source"""