+++ packages/orq-rc/src/orq_ai_sdk/models/listidentitiesop.py
@@ -208,7 +208,7 @@ class ListIdentitiesData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-04-23T07:35:28.002Z")
+    updated: Optional[datetime] = parse_datetime("2026-04-23T12:18:03.924Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")