+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPWM88465ZGPKYNXMTS9TY7J"
+        "01KPX4DPK2BMMG31RCPGVPME47"
     )
     r"""The unique identifier of the data source"""