+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPWM88493X5G9RJ7JF2K9JJR"
+        "01KPX4DPK6446MVDNQ1K65HZSN"
     )
     r"""The unique identifier of the data source"""