+++ packages/orq-rc/RELEASES.md
@@ -7289,3 +7289,13 @@ Based on:
 - [python v4.8.0-rc.52] packages/orq-rc
 ### Releases
 - [PyPI v4.8.0-rc.52] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.52 - packages/orq-rc
+
+## 2026-04-23 12:21:08
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.8.0-rc.53] packages/orq-rc
+### Releases
+- [PyPI v4.8.0-rc.53] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.53 - packages/orq-rc